refactor: colocate utility and history feature slices
