+++ src/content/main/features/action-recorder/action-recorder-dialog.styles.js
