+++ src/content/main/features/action-recorder/action-recorder.js
