+++ src/content/main/features/marathon-export/export-progress-dialog.js
@@ -1,5 +1,8 @@
 import { LitElement, html, nothing } from 'lit';
-import { componentFoundationStyles, dialogFoundationStyles } from './styles/foundations.js';
+import {
+    componentFoundationStyles,
+    dialogFoundationStyles
+} from '../../../../components/styles/foundations.js';
 import { exportProgressDialogStyles } from './export-progress-dialog.styles.js';
 
 const EXPORT_PROGRESS_TAG = 'edvibe-toolbox-export-progress';
