+++ src/content/main/features/reset-lessons/reset-lessons-dialog.js
@@ -1,5 +1,8 @@
 import { LitElement, html, nothing } from 'lit';
-import { componentFoundationStyles, dialogFoundationStyles } from './styles/foundations.js';
+import {
+    componentFoundationStyles,
+    dialogFoundationStyles
+} from '../../../../components/styles/foundations.js';
 import { resetLessonsDialogStyles } from './reset-lessons-dialog.styles.js';
 
 const RESET_DIALOG_TAG = 'edvibe-toolbox-reset-dialog';
