+++ src/content/main/features/reset-lessons/reset-lessons-dialog.styles.js
