+++ src/content/main/index.js
@@ -13,14 +13,14 @@ import * as historyRetentionApi from '../../shared/execution-history-retention.j
 import * as historyExportApi from '../../shared/execution-history-export.js';
 import * as storageBridgeApi from '../../shared/chrome-storage-bridge.js';
 import * as historyServiceApi from '../../shared/execution-history-service.js';
-import '@/components/export-progress-dialog.js';
-import '@/components/reset-lessons-dialog.js';
-import * as historyDialogApi from '../../components/execution-history-dialog.js';
-import * as historyFeatureApi from '../../features/execution-history.js';
-import * as exportApi from '../../features/marathon-export.js';
-import * as resetApi from '../../features/reset-lessons.js';
-import * as recorderApi from '../../features/action-recorder.js';
-import * as recorderDialogApi from '../../components/action-recorder-dialog.js';
+import './features/marathon-export/export-progress-dialog.js';
+import './features/reset-lessons/reset-lessons-dialog.js';
+import * as historyDialogApi from './features/execution-history/execution-history-dialog.js';
+import * as historyFeatureApi from './features/execution-history/execution-history.js';
+import * as exportApi from './features/marathon-export/marathon-export.js';
+import * as resetApi from './features/reset-lessons/reset-lessons.js';
+import * as recorderApi from './features/action-recorder/action-recorder.js';
+import * as recorderDialogApi from './features/action-recorder/action-recorder-dialog.js';
 import * as batchAccessApi from '../../features/batch-lesson-access.js';
 import * as batchAccessHistoryApi from '../../features/batch-lesson-access-history.js';
 import * as batchAccessDialogApi from '../../components/batch-lesson-access-dialog.js';
